refactor: Delete custom recursive-task type system from Daytona provider
Remove result_types.py (563 lines) and result_views.py (319 lines) which
implemented a parallel orchestration layer (RecursiveTaskSpec, AgentNode,
ChildLink, FinalArtifact, RolloutSummary, etc.) that duplicates what
dspy.RLM already provides internally.

Slim types.py from 442 to 283 lines by removing PromptHandle,
PromptSliceRef, PromptManifest, ContextSource prompt types, and
RolloutBudget. Move DaytonaSmokeResult into types.py with to_dict().

Clean up __init__.py exports and state.py imports.

Phase 1.1 of simplify-rlm-architecture plan.
