+++ .gitignore
@@ -119,3 +119,4 @@ src/frontend/test-results/.last-run.json
 .posthog-events.json
 src/frontend/posthog-setup-report.md
 PLANS.md
+.output.txt
