+++ src/fleet_rlm/integrations/providers/daytona/__init__.py
@@ -13,51 +13,25 @@ from .runtime import (
 )
 from .smoke import run_daytona_smoke
 from .types import (
-    AgentNode,
-    ChildLink,
-    ChildTaskResult,
     ContextSource,
-    DaytonaEvidenceRef,
-    DaytonaRunResult,
-    DaytonaSmokeResult,
-    ExecutionObservation,
-    FinalArtifact,
-    PromptHandle,
-    PromptManifest,
-    PromptSliceRef,
-    RecursiveTaskSpec,
-    RolloutBudget,
-    RolloutSummary,
+    DaytonaRunCancelled,
+    SandboxLmRuntimeConfig,
     SandboxSpec,
-    TaskSourceProvenance,
 )
 
 __all__ = [
-    "AgentNode",
-    "ChildLink",
-    "ChildTaskResult",
     "ContextSource",
     "DAYTONA_PERSISTENT_VOLUME_MOUNT_PATH",
     "DaytonaConfigError",
     "DaytonaDiagnosticError",
-    "DaytonaEvidenceRef",
     "DaytonaInterpreter",
-    "DaytonaRunResult",
+    "DaytonaRunCancelled",
     "DaytonaSandboxRuntime",
     "DaytonaSandboxSession",
-    "DaytonaSmokeResult",
     "DaytonaWorkbenchChatAgent",
-    "ExecutionObservation",
-    "FinalArtifact",
-    "PromptHandle",
-    "PromptManifest",
-    "PromptSliceRef",
-    "RecursiveTaskSpec",
     "ResolvedDaytonaConfig",
-    "RolloutBudget",
-    "RolloutSummary",
+    "SandboxLmRuntimeConfig",
     "SandboxSpec",
-    "TaskSourceProvenance",
     "resolve_daytona_config",
     "run_daytona_smoke",
 ]
